backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn==23.0.0
